requirements.txt
google-api-python-client
python-dotenv
youtube-transcript-api
anthropic
markdown
ebooklib
requests
